% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_miner.R
\name{alpha_miner}
\alias{alpha_miner}
\title{Alpha Miner}
\usage{
alpha_miner(eventlog, flavor = c("traditional"))
}
\arguments{
\item{eventlog}{Event log}

\item{flavor}{Algorithm flavor, currently only the traditional algoritm}
}
\value{
A Petri Net
}
\description{
Alpha Miner
}
