% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cj_utils.R
\name{BP}
\alias{BP}
\title{Calculate inter-class inertia}
\usage{
BP(DF, clusters)
}
\arguments{
\item{clusters}{a vector specifying the cluster of each individual.}

\item{data}{a data.frame object.}
}
\value{
numeric
}
\description{
Calculate inter-class inertia
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
