% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_utils.R
\name{calc.centros}
\alias{calc.centros}
\title{Calculation of the center of clusters}
\usage{
calc.centros(data, clusters)
}
\arguments{
\item{data}{a data.frame object.}

\item{clusters}{a vector specifying the cluster of each individual.}
}
\value{
list
}
\description{
Calculation of the center of clusters
}
\examples{
clusters <- factor(kmeans(iris[, -5], 3)$cluster)
calc.centros(iris[, -5], clusters)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
