% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos_utils.R
\name{datos.disyuntivos}
\alias{datos.disyuntivos}
\title{Create disjunctive columns to a data.frame.}
\usage{
datos.disyuntivos(data, var)
}
\arguments{
\item{data}{a data.frame object.}

\item{var}{the column name to apply disjunctive code.}
}
\value{
data.frame
}
\description{
Create disjunctive columns to a data.frame.
}
\examples{
datos.disyuntivos(iris, "Species")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
