% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal_utils.R
\name{dfnormal}
\alias{dfnormal}
\title{Normal data.frame}
\usage{
dfnormal(data)
}
\arguments{
\item{data}{a data.frame object only with the numeric columns.}
}
\value{
data.frame
}
\description{
Normal data.frame
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
