% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoveR.R
\docType{package}
\name{discoveR}
\alias{discoveR-package}
\alias{discoveR}
\title{Exploratory Data Analysis System}
\description{
Performs an exploratory data analysis through a 'shiny' interface. 
It includes basic methods such as the mean, median, mode, normality test, 
among others. It also includes clustering techniques such as Principal 
Components Analysis, Hierarchical Clustering and the K-Means Method.
}
\details{
\tabular{ll}{
Package: \tab discoveR\cr
Type: \tab Package\cr
Version: \tab 3.1.7\cr
Date: \tab 2025-02-12\cr
License: \tab GPL (>=2)\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://promidat.website/}
  \item \url{https://github.com/PROMiDAT/discoveR}
  \item Report bugs at \url{https://github.com/PROMiDAT/discoveR/issues}
}

}
\author{
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>\cr
\itemize{
  \item Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
  \item Diego Jiménez Alvarado
}
}
\keyword{package}
