% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_afcm_utils.R
\name{e_afcmvar}
\alias{e_afcmvar}
\title{AFCM plot of variables}
\usage{
e_afcmvar(modelo, axes = c(1, 2), colorVar = "forestgreen")
}
\arguments{
\item{modelo}{an object of class AFCM [FactoMineR].}

\item{axes}{a numeric vector of length 2 specifying the dimensions to be plotted.}

\item{colorVar}{a color for the variables.}
}
\value{
echarts4r plot
}
\description{
AFCM plot of variables
}
\examples{
data("poison", package = "FactoMineR")
poison.active <- poison[1:55, 5:15]
p <- FactoMineR::MCA(poison.active, graph = FALSE)
e_afcmvar(p)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
