% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_afc_utils.R
\name{e_balloon}
\alias{e_balloon}
\title{Balloonplot}
\usage{
e_balloon(datos)
}
\arguments{
\item{datos}{a data frame object.}
}
\value{
echarts4r plot
}
\description{
Balloonplot
}
\examples{
e_balloon(iris)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
