% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_utils.R
\name{e_cat}
\alias{e_cat}
\title{Barplot for categoric variable by clusters.}
\usage{
e_cat(clusters, var, colores = NULL, escalar = T)
}
\arguments{
\item{clusters}{a vector specifying the cluster of each individual.}

\item{var}{a factor column of a data.frame.}

\item{colores}{a vector of color for each cluster.}

\item{escalar}{a boolean value specifying if use percentage or real values.}
}
\value{
echarts4r plot
}
\description{
Barplot for categoric variable by clusters.
}
\examples{
clusters <- factor(kmeans(iris[, -5], 3)$cluster)
e_cat(clusters, iris[, 5], colores = c("steelblue", "pink", "forestgreen"))

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
