% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_correlacion_utils.R
\name{e_cor}
\alias{e_cor}
\title{Correlation plot}
\usage{
e_cor(x, colors = c("#FF5733", "#F8F5F5", "#2E86C1"))
}
\arguments{
\item{x}{a data.frame with correlation values.}

\item{colors}{a vector of lenght 3 with color values.}
}
\value{
echarts4r plot
}
\description{
Correlation plot
}
\examples{
p <- round(cor(iris[, -5]), 3)
e_cor(p)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
