% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distribuciones_utils.R
\name{e_hboxplot}
\alias{e_hboxplot}
\title{Generate horizontal boxplot}
\usage{
e_hboxplot(e, serie, name = NULL, outliers = TRUE, ...)
}
\description{
Generate horizontal boxplot
}
\keyword{internal}
