% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distribuciones_utils.R
\name{e_histboxplot}
\alias{e_histboxplot}
\title{Histogram + boxplot}
\usage{
e_histboxplot(
  data,
  var.name,
  colorBar = "steelblue",
  colorPoint = "red",
  titulos = c("Minimo", "Primer Cuartil", "Mediana", "Tercer Cuartil", "Maximo")
)
}
\arguments{
\item{data}{a numeric column of a data.frame.}

\item{var.name}{a character value specifying the name of the variable.}

\item{colorBar}{a color for the bars.}

\item{colorPoint}{a color for the points.}

\item{titulos}{a character vector of length 5 specifying the titles to use on legend.}
}
\value{
echarts4r plot
}
\description{
Histogram + boxplot
}
\examples{
e_histboxplot(iris$Sepal.Width, "Sepal.Width")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
