% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal_utils.R
\name{e_histnormal}
\alias{e_histnormal}
\title{Normal plot}
\usage{
e_histnormal(
  data,
  colorbar = "steelblue",
  colorline = "gray",
  nombres = c("Histograma", "Curva Normal")
)
}
\arguments{
\item{data}{a numeric column of a data.frame.}

\item{colorbar}{a color for the bars.}

\item{colorline}{a color for the line.}

\item{nombres}{a character vector of length 2 specifying the titles to use on legend.}
}
\value{
echarts4r plot
}
\description{
Normal plot
}
\examples{
e_histnormal(iris$Sepal.Length)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
