% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_utils.R
\name{e_inercia}
\alias{e_inercia}
\title{Inertia plot of clusterization}
\usage{
e_inercia(
  data,
  titulos = c("Inercia", "Inercia Inter-Clase", "Inercia Inter-Clase")
)
}
\arguments{
\item{data}{a data.frame object with the inertia values.}

\item{titulos}{a character vector of length 3 specifying the titles to use on legend.}
}
\value{
echarts4r plot
}
\description{
Inertia plot of clusterization
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
