% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kmedias_utils.R
\name{e_jambu}
\alias{e_jambu}
\title{Jambu Elbow plot}
\usage{
e_jambu(data, max.clusters)
}
\arguments{
\item{data}{a data.frame object.}

\item{max.clusters}{a numeric value specifying the number of times to generate the model.}
}
\value{
echarts4r plot
}
\description{
Jambu Elbow plot
}
\examples{
e_jambu(iris[, -5], 10)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
