% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal_utils.R
\name{e_qq}
\alias{e_qq}
\title{Qplot + Qline}
\usage{
e_qq(data, colorpoint = "steelblue", colorline = "gray")
}
\arguments{
\item{data}{a numeric column of a data.frame.}

\item{colorpoint}{a color for the points.}

\item{colorline}{a color for the line.}
}
\value{
echarts4r plot
}
\description{
Qplot + Qline
}
\examples{
e_qq(iris$Sepal.Length)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
