% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cj_utils.R
\name{gg_dendrograma}
\alias{gg_dendrograma}
\title{Dendrogram plot}
\usage{
gg_dendrograma(model, k, colors = NULL)
}
\arguments{
\item{model}{an object of class hclust.}

\item{k}{a vector specifying the cluster of each individual.}

\item{colors}{a vector of color for each cluster.}
}
\value{
ggplot
}
\description{
Dendrogram plot
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
