% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_utils.R
\name{hc_inercia}
\alias{hc_inercia}
\title{Inertia plot of clusterization}
\usage{
hc_inercia(
  data,
  nombre.archivo,
  titulos = c("Inercia", "Inercia Inter-Clase", "Inercia Inter-Clase")
)
}
\arguments{
\item{data}{a data.frame object with the inertia values.}

\item{nombre.archivo}{a character value specifying the name to use when the plot is downloaded.}

\item{titulos}{a character vector of length 3 specifying the titles to use on legend.}
}
\value{
Highchart plot
}
\description{
Inertia plot of clusterization
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
