% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kmedias_utils.R
\name{hc_jambu}
\alias{hc_jambu}
\title{Jambu Elbow plot}
\usage{
hc_jambu(data, max.clusters, nombre.archivo = NULL)
}
\arguments{
\item{data}{a data.frame object.}

\item{max.clusters}{a numeric value specifying the number of times to generate the model.}

\item{nombrearchivo}{a character value specifying the name to use when the plot is downloaded.}
}
\value{
Highchart plot
}
\description{
Jambu Elbow plot
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
