% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_utils.R
\name{hc_radar}
\alias{hc_radar}
\title{Radar representation for centers of clusters.}
\usage{
hc_radar(centros, nombre.archivo, colores = NULL)
}
\arguments{
\item{centros}{a data.frame object with the centers of the clusters.}

\item{colores}{a vector of color for each cluster.}

\item{nombrearchivo}{a character value specifying the name to use when the plot is downloaded.}
}
\value{
Highchart plot
}
\description{
Radar representation for centers of clusters.
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
