% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kmedias_utils.R
\name{hc_silhouette}
\alias{hc_silhouette}
\title{Silhouette plot}
\usage{
hc_silhouette(data, max.clusters)
}
\arguments{
\item{data}{a data.frame object.}

\item{max.clusters}{a numeric value specifying the number of times to generate the model.}
}
\value{
Highchart plot
}
\description{
Silhouette plot
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
