% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distribuciones_utils.R
\name{hchistboxplot}
\alias{hchistboxplot}
\title{Histogram + boxplot}
\usage{
hchistboxplot(
  data,
  nombrearchivo = NULL,
  colorBar = "steelblue",
  colorPoint = "red",
  outlier.name = "",
  titulos = c("Minimo", "Primer Cuartil", "Mediana", "Tercer Cuartil", "Maximo")
)
}
\arguments{
\item{data}{a numeric column of a data.frame.}

\item{nombrearchivo}{a character value specifying the name to use when the plot is downloaded.}

\item{colorBar}{a color for the bars.}

\item{colorPoint}{a color for the points.}

\item{outlier.name}{a character value specifying the name to use on the tooltip of points.}

\item{titulos}{a character vector of length 5 specifying the titles to use on legend.}
}
\value{
Highchart plot
}
\description{
Histogram + boxplot
}
\examples{
hchistboxplot(iris$Sepal.Width)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
