% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_acp_utils.R
\name{hcpcaind}
\alias{hcpcaind}
\title{PCA plot of individuals}
\usage{
hcpcaind(
  modelo,
  axes = c(1, 2),
  nombrearchivo = NULL,
  colorInd = "steelblue",
  cos2 = 0,
  colorCos = "firebrick",
  titulos = c("Bien Representados", "Mal Representados")
)
}
\arguments{
\item{modelo}{an object of class PCA [FactoMineR].}

\item{axes}{a numeric vector of length 2 specifying the dimensions to be plotted.}

\item{nombrearchivo}{a character value specifying the name to use when the plot is downloaded.}

\item{colorInd}{a color for the individuals well represented.}

\item{cos2}{a numeric value from 0 to 1 specifying the quality of the individuals.}

\item{colorCos}{a color for individuals badly represented.}

\item{titulos}{a character vector of length 2 specifying the titles to use on legend.}
}
\value{
Highchart plot
}
\description{
PCA plot of individuals
}
\examples{
p <- discoveR:::PCA(iris[, -5], graph = FALSE)
hcpcaind(p)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
