% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal_utils.R
\name{hcqq}
\alias{hcqq}
\title{Qplot + Qline}
\usage{
hcqq(data, nombrearchivo = NULL, colorpoint = "steelblue", colorline = "gray")
}
\arguments{
\item{data}{a numeric column of a data.frame.}

\item{nombrearchivo}{a character value specifying the name to use when the plot is downloaded.}

\item{colorpoint}{a color for the points.}

\item{colorline}{a color for the line.}
}
\value{
Highchart plot
}
\description{
Qplot + Qline
}
\examples{
hcqq(iris$Sepal.Length)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
