% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cj_utils.R
\name{inercia.total}
\alias{inercia.total}
\title{Calculate total inertia}
\usage{
inercia.total(DF)
}
\arguments{
\item{DF}{a data.frame object.}
}
\value{
numeric
}
\description{
Calculate total inertia
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
