% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_discover.R
\name{init_discover}
\alias{init_discover}
\title{This function will start DiscoveR}
\usage{
init_discover()
}
\value{
Nothing
}
\description{
An interactive Shiny application for exploring data.
}
\details{
Start DiscoveR

This starts the DiscoveR application on the user's local computer.
}
\examples{
\dontrun{
 if(interactive()){
   init_discover()
 }
}
}
\keyword{DiscoveR}
