% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_acp.R
\name{mod_acp_server}
\alias{mod_acp_server}
\title{acp Server Function}
\usage{
mod_acp_server(input, output, session, updateData)
}
\description{
acp Server Function
}
\keyword{(Biplot)}
\keyword{(Resultados}
\keyword{(individuals)}
\keyword{(variables)}
\keyword{2D}
\keyword{3D}
\keyword{Choose}
\keyword{Generate}
\keyword{K-medias}
\keyword{PCA}
\keyword{Plot}
\keyword{data}
\keyword{internal}
\keyword{load}
\keyword{numéricos)}
\keyword{on}
\keyword{or}
\keyword{plot}
