% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cj.R
\name{mod_cj_server}
\alias{mod_cj_server}
\title{cj Server Function}
\usage{
mod_cj_server(input, output, session, updateData)
}
\description{
cj Server Function
}
\keyword{(CJ)}
\keyword{(Categórico)}
\keyword{(Cluster}
\keyword{(Horizontal)}
\keyword{(Radar)}
\keyword{(Resultados}
\keyword{(Vertical)}
\keyword{2D}
\keyword{3D}
\keyword{Agregar}
\keyword{Choose}
\keyword{Clusteres}
\keyword{Colores}
\keyword{Generate}
\keyword{Inertia}
\keyword{K-medias}
\keyword{Mapa}
\keyword{Mostrar}
\keyword{Plot}
\keyword{Update}
\keyword{a}
\keyword{data}
\keyword{datos}
\keyword{de}
\keyword{dendrogram}
\keyword{hclust}
\keyword{internal}
\keyword{jerarquico)}
\keyword{load}
\keyword{numéricos)}
\keyword{on}
\keyword{or}
\keyword{plot}
\keyword{tabla}
