% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_correlacion.R
\name{mod_correlacion_server}
\alias{mod_correlacion_server}
\title{correlacion Server Function}
\usage{
mod_correlacion_server(input, output, session, updateData)
}
\description{
correlacion Server Function
}
\keyword{Correlaciones}
\keyword{Gráfico}
\keyword{Resultados}
\keyword{de}
\keyword{internal}
\keyword{numéricos}
