% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dispersion.R
\name{mod_dispersion_server}
\alias{mod_dispersion_server}
\title{dispersion Server Function}
\usage{
mod_dispersion_server(input, output, session, updateData)
}
\description{
dispersion Server Function
}
\keyword{Plot}
\keyword{Scatter}
\keyword{Update}
\keyword{data}
\keyword{internal}
\keyword{load}
\keyword{on}
