% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distribuciones.R
\name{mod_distribuciones_server}
\alias{mod_distribuciones_server}
\title{distribuciones Server Function}
\usage{
mod_distribuciones_server(input, output, session, updateData)
}
\description{
distribuciones Server Function
}
\keyword{(Categóricas)}
\keyword{(Númericas)}
\keyword{Distribuciones}
\keyword{Gráfico}
\keyword{Update}
\keyword{data}
\keyword{de}
\keyword{internal}
\keyword{load}
\keyword{on}
