% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal.R
\name{mod_normal_server}
\alias{mod_normal_server}
\title{normal Server Function}
\usage{
mod_normal_server(input, output, session, updateData)
}
\description{
normal Server Function
}
\keyword{+}
\keyword{Grafico}
\keyword{Resumen}
\keyword{Test}
\keyword{Update}
\keyword{data}
\keyword{de}
\keyword{internal}
\keyword{load}
\keyword{normalidad}
\keyword{on}
\keyword{qqline}
\keyword{qqplot}
