% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_numerico.R
\name{mod_r_numerico_server}
\alias{mod_r_numerico_server}
\title{r_numerico Server Function}
\usage{
mod_r_numerico_server(input, output, session, updateData)
}
\description{
r_numerico Server Function
}
\keyword{Resumen}
\keyword{Update}
\keyword{data}
\keyword{internal}
\keyword{load}
\keyword{numérico}
\keyword{on}
