% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_utils.R
\name{plotly_mapa}
\alias{plotly_mapa}
\title{PCA plot of individuals colored by clusters}
\usage{
plotly_mapa(pca.model, clusters, colores = NULL, ejes = c(1, 2, 3))
}
\arguments{
\item{pca.model}{an object of class PCA [FactoMineR].}

\item{clusters}{a vector specifying the cluster of each individual.}

\item{colores}{a vector of color for each cluster.}

\item{ejes}{a numeric vector of length 3 specifying the dimensions to be plotted.}
}
\value{
plotly plot
}
\description{
PCA plot of individuals colored by clusters
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
