% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_numerico_utils.R
\name{resumen.categorico}
\alias{resumen.categorico}
\title{Funciones Resumen Categórico}
\usage{
resumen.categorico(data, variable)
}
\value{
functions
}
\description{
Funciones Resumen Categórico
}
\author{
Diego
}
