% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_numerico_utils.R
\name{resumen.numerico}
\alias{resumen.numerico}
\title{Funciones Resumen Numérico}
\usage{
resumen.numerico(data, variable, idioma = "es")
}
\value{
functions
}
\description{
Funciones Resumen Numérico
}
\author{
Diego
}
