% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos_utils.R
\name{var.categoricas}
\alias{var.categoricas}
\title{Filter category variables of a data.frame}
\usage{
var.categoricas(data)
}
\arguments{
\item{data}{a data.frame object.}
}
\value{
data.frame
}
\description{
Filter category variables of a data.frame
}
\examples{
var.categoricas(iris)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
