% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos_utils.R
\name{var.numericas}
\alias{var.numericas}
\title{Filter numeric variables of a data.frame}
\usage{
var.numericas(data)
}
\arguments{
\item{data}{a data.frame object.}
}
\value{
data.frame
}
\description{
Filter numeric variables of a data.frame
}
\examples{
var.numericas(iris)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
