% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cj_utils.R
\name{BP}
\alias{BP}
\title{Calculate inter-class inertia}
\usage{
BP(DF, clusters)
}
\arguments{
\item{DF}{a data.frame object.}

\item{clusters}{a vector specifying the cluster of each individual.}
}
\value{
numeric
}
\description{
Calculate inter-class inertia
}
\examples{
m <- hclust(dist(iris[, -5]))
BP(iris[, -5], cutree(m, 3))

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
