% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal_utils.R
\name{dfnormal}
\alias{dfnormal}
\title{Data.frame with normal test}
\usage{
dfnormal(data)
}
\arguments{
\item{data}{a data.frame object only with the numeric columns.}
}
\value{
data.frame
}
\description{
Data.frame with normal test
}
\examples{
dfnormal(iris[, -5])

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
