% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoveR.R
\docType{package}
\name{discoveR}
\alias{discoveR}
\title{Exploratory Data Analysis System}
\description{
Performs an exploratory data analysis through a 'shiny' interface. 
It includes basic methods such as the mean, median, mode, normality test, 
among others. It also includes clustering techniques such as Principal 
Components Analysis, Hierarchical Clustering and the K-Means Method.
}
\details{
\tabular{ll}{
Package: \tab discoveR\cr
Type: \tab Package\cr
Version: \tab 2.2.6\cr
Date: \tab 2021-10-08\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Oldemar Rodriguez Rojas \cr
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
}
\keyword{package}
