% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_utils.R
\name{hc_cat}
\alias{hc_cat}
\title{Barplot for categoric variable by clusters.}
\usage{
hc_cat(clusters, var, nombre.archivo = NULL, colores = NULL, escalar = T)
}
\arguments{
\item{clusters}{a vector specifying the cluster of each individual.}

\item{var}{a factor column of a data.frame.}

\item{colores}{a vector of color for each cluster.}

\item{escalar}{a boolean value specifying if use percentage or real values.}

\item{nombrearchivo}{a character value specifying the name to use when the plot is downloaded.}
}
\value{
Highchart plot
}
\description{
Barplot for categoric variable by clusters.
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
