% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_utils.R
\name{hc_mapa}
\alias{hc_mapa}
\title{PCA plot of individuals colored by clusters}
\usage{
hc_mapa(
  pca.model,
  clusters,
  nombre.archivo = NULL,
  colores = NULL,
  ejes = c(1, 2)
)
}
\arguments{
\item{pca.model}{an object of class PCA [FactoMineR].}

\item{clusters}{a vector specifying the cluster of each individual.}

\item{colores}{a vector of color for each cluster.}

\item{ejes}{a numeric vector of length 2 specifying the dimensions to be plotted.}

\item{nombrearchivo}{a character value specifying the name to use when the plot is downloaded.}
}
\value{
Highchart plot
}
\description{
PCA plot of individuals colored by clusters
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
