% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kmedias_utils.R
\name{hc_silhouette}
\alias{hc_silhouette}
\title{Silhouette plot}
\usage{
hc_silhouette(data, max.clusters, nombre.archivo = NULL)
}
\arguments{
\item{data}{a data.frame object.}

\item{max.clusters}{a numeric value specifying the number of times to generate the model.}

\item{nombre.archivo}{a character value specifying the name to use when the plot is downloaded.}
}
\value{
Highchart plot
}
\description{
Silhouette plot
}
\examples{
hc_silhouette(iris[, -5], 10)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
