% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_acp.R
\name{mod_acp_server}
\alias{mod_acp_server}
\title{acp Server Function}
\usage{
mod_acp_server(input, output, session, updateData)
}
\description{
acp Server Function
}
\keyword{(Biplot)}
\keyword{(Resultados}
\keyword{(individuals)}
\keyword{(variables)}
\keyword{2D}
\keyword{3D}
\keyword{ACP}
\keyword{Choose}
\keyword{Componentes}
\keyword{Contribución}
\keyword{Correlación}
\keyword{Cosenos}
\keyword{Cuadrados}
\keyword{Eje}
\keyword{Explicada}
\keyword{Generate}
\keyword{Individuos}
\keyword{Modelo}
\keyword{PCA}
\keyword{Plot}
\keyword{Variables}
\keyword{Varianza}
\keyword{cada}
\keyword{con}
\keyword{data}
\keyword{de}
\keyword{internal}
\keyword{las}
\keyword{load}
\keyword{los}
\keyword{numéricos)}
\keyword{on}
\keyword{or}
\keyword{plot}
\keyword{por}
\keyword{variables}
