% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cj_utils.R
\name{WP}
\alias{WP}
\title{Calculate intra-class inertia}
\usage{
WP(DF, clusters)
}
\arguments{
\item{clusters}{a vector specifying the cluster of each individual.}

\item{data}{a data.frame object.}
}
\value{
numeric
}
\description{
Calculate intra-class inertia
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
